refactor(tests): Make lock manager test a reusable TCK
