+++ kork-jedis/kork-jedis.gradle
@@ -10,6 +10,7 @@ dependencies {
   api "com.netflix.spectator:spectator-api"
 
   testImplementation "com.hubspot.jinjava:jinjava"
+  testImplementation project(":kork-core-tck")
   testImplementation project(":kork-jedis-test")
   testImplementation "org.codehaus.groovy:groovy-all"
   testImplementation "org.spockframework:spock-core"
