+++ kork-jedis/src/test/groovy/com/netflix/spinnaker/kork/jedis/locking/RedisLockManagerSpec.groovy
@@ -16,48 +16,38 @@
 
 package com.netflix.spinnaker.kork.jedis.locking
 
+import java.util.function.Consumer
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.netflix.spectator.api.NoopRegistry
-
 import com.netflix.spinnaker.kork.jedis.EmbeddedRedis
 import com.netflix.spinnaker.kork.jedis.JedisClientDelegate
 import com.netflix.spinnaker.kork.jedis.lock.RedisLockManager
+import com.netflix.spinnaker.kork.lock.BaseLockManagerSpec
 import com.netflix.spinnaker.kork.lock.LockManager
 import redis.clients.jedis.JedisCommands
 import redis.clients.jedis.JedisPool
 import spock.lang.Shared
-import spock.lang.Specification
 
-import java.time.Clock
-import java.time.Duration
-import java.util.concurrent.Callable
-import java.util.concurrent.atomic.AtomicInteger
-import java.util.function.Consumer
-
-import static com.netflix.spinnaker.kork.lock.LockManager.LockStatus.*
-import static com.netflix.spinnaker.kork.lock.RefreshableLockManager.*
-
-import static com.netflix.spinnaker.kork.lock.LockManager.*
-
-
-class RedisLockManagerSpec extends Specification {
+class RedisLockManagerSpec extends BaseLockManagerSpec {
   @Shared def embeddedRedis = EmbeddedRedis.embed()
   def jedisPool = embeddedRedis.getPool() as JedisPool
   def objectMapper = new ObjectMapper()
-  def clock = Clock.systemDefaultZone()
   def registry = new NoopRegistry()
   def redisClientDelegate = new JedisClientDelegate(jedisPool)
   def heartbeatRateMillis = 30L
-  def testLockMaxDurationMillis = 1000L
-  def redisLockManager = new RedisLockManager(
-    "testOwner",
-    clock,
-    registry,
-    objectMapper,
-    redisClientDelegate,
-    Optional.of(heartbeatRateMillis),
-    Optional.of(testLockMaxDurationMillis)
-  )
+
+  @Override
+  protected LockManager subject() {
+    return new RedisLockManager(
+      "testOwner",
+      clock,
+      registry,
+      objectMapper,
+      redisClientDelegate,
+      Optional.of(heartbeatRateMillis),
+      Optional.of(testLockMaxDurationMillis)
+    )
+  }
 
   def setup() {
     jedisPool.resource.withCloseable {
@@ -69,182 +59,31 @@ class RedisLockManagerSpec extends Specification {
     embeddedRedis.destroy()
   }
 
-  def "should acquire a simple lock and auto release"() {
-    when:
-    def result = redisLockManager.acquireLock("veryImportantLock", testLockMaxDurationMillis, {
-      return "Run after lock is safely acquired."
-    } as Callable<String>)
-
-    then:
-    result.lockStatus == ACQUIRED
-    result.released == true
-    result.onLockAcquiredCallbackResult == "Run after lock is safely acquired."
-  }
-
-  def "should store arbitrary data as attributes alongside the lock"() {
-    given:
-    def lockOptions = new LockOptions()
-      .withMaximumLockDuration(Duration.ofMillis(testLockMaxDurationMillis))
-      .withLockName("veryImportantLock")
-      .withAttributes(["key:value","key2:value2"])
-
-    when:
-    def result = redisLockManager.acquireLock(lockOptions, {
-      return "Lock with data in attributes"
-    } as Callable<String>)
-
-    then:
-    result.lock.attributes == "key:value;key2:value2"
-  }
-
-  def "should fail to acquire an already taken lock"() {
-    given:
-    def lockOptions = new LockOptions()
-      .withMaximumLockDuration(Duration.ofMillis(testLockMaxDurationMillis))
-      .withLockName("veryImportantLock")
-
-    and:
-    redisLockManager.tryCreateLock(lockOptions)
-
-    when:
-    def result = redisLockManager.acquireLock(lockOptions, {
-      return "attempting to acquire lock"
-    } as Callable<String>)
-
-    then:
-    result.lockStatus == TAKEN
-    result.onLockAcquiredCallbackResult == null
-  }
-
-  def "should acquire with heartbeat"() {
-    given:
-    def lockOptions = new LockOptions()
-      .withMaximumLockDuration(Duration.ofMillis(testLockMaxDurationMillis))
-      .withLockName("veryImportantLock")
-
-    when:
-    def result = redisLockManager.acquireLock(lockOptions, {
-      // simulates long running task
-      Thread.sleep(100)
-      "Done"
-    } as Callable<String>)
-
-    then:
-    result.released == true
-    result.lockStatus == ACQUIRED
-    result.onLockAcquiredCallbackResult == "Done"
-
-    when:
-    result = redisLockManager.acquireLock(lockOptions.withLockName("withRunnable"), {
-      // simulates long running task
-      Thread.sleep(100)
-    } as Runnable)
-
-    then:
-    result.onLockAcquiredCallbackResult == null
-    result.lockStatus == ACQUIRED
-  }
-
-  def "should propagate exception on callback failure"() {
-    given:
-    def lockName = "veryImportantLock"
-    def onLockAcquiredCallback = {
-      throw new IllegalStateException("Failure")
-    }
-
-    when:
-    redisLockManager.acquireLock(lockName, testLockMaxDurationMillis, onLockAcquiredCallback)
-
-    then:
-    thrown(LockManager.LockCallbackException)
-  }
-
-  def "should release a lock"() {
-    given:
-    def lockOptions = new LockManager.LockOptions()
-      .withMaximumLockDuration(Duration.ofMillis(testLockMaxDurationMillis))
-      .withLockName("veryImportantLock")
-
-    and:
-    def lock = redisLockManager.tryCreateLock(lockOptions)
-
+  @Override protected void ensureLockExists(String lockName) {
     redisClientDelegate.withCommandsClient({ JedisCommands c ->
       // ensure that the lock was actually created
-      assert c.exists("{korkLock:veryimportantlock}")
+      assert c.exists("{korkLock:${lockName.toLowerCase()}}")
     } as Consumer<JedisCommands>)
+  }
 
-    when:
-    def response = redisLockManager.tryReleaseLock(lock, true)
-
-    then:
-    response == LockReleaseStatus.SUCCESS.toString()
-
+  @Override protected void ensureLockReleased(String lockName) {
     redisClientDelegate.withCommandsClient({ JedisCommands c ->
       // ensure that the lock no longer exists (default `successInterval` of zero should immediately expire the lock)
-      assert !c.exists("{korkLock:veryimportantlock}")
+      assert !c.exists("{korkLock:${lockName.toLowerCase()}}")
     } as Consumer<JedisCommands>)
   }
 
-  def "should heartbeat by updating lock ttl"() {
-    given:
-    def heartbeatRetriesOnFailure = new AtomicInteger(1)
-    def lockOptions = new LockOptions()
-      .withMaximumLockDuration(Duration.ofMillis(testLockMaxDurationMillis))
-      .withLockName("veryImportantLock")
-
-    and:
-    def lock = redisLockManager.tryCreateLock(lockOptions)
-    def request = new HeartbeatLockRequest(lock, heartbeatRetriesOnFailure, clock, Duration.ofMillis(200), false)
-    Thread.sleep(10)
-
-    when:
-    def response = redisLockManager.heartbeat(request)
-    Thread.sleep(10)
-
-    then:
-    response.lockStatus == LockHeartbeatStatus.SUCCESS
-
-    when: "Late heartbeat resulting in expired lock "
-    request = new HeartbeatLockRequest(lock, heartbeatRetriesOnFailure, clock, Duration.ofMillis(200), false)
-    redisLockManager.heartbeat(request)
-    Thread.sleep(30)
-    response = redisLockManager.heartbeat(request)
-
-    then:
-    response.lockStatus == LockHeartbeatStatus.EXPIRED
-  }
-
-  def "should support success and failure intervals when releasing a lock"() {
-    given:
-    def lockOptions = new LockOptions()
-        .withMaximumLockDuration(Duration.ofMillis(testLockMaxDurationMillis))
-        .withLockName("veryImportantLock")
-        .withSuccessInterval(Duration.ofSeconds(100))
-        .withFailureInterval(Duration.ofSeconds(25))
-
-    and:
-    def lock = redisLockManager.tryCreateLock(lockOptions)
-
-    when:
-    def response = redisLockManager.tryReleaseLock(lock, true)
-
-    then:
-    response == LockReleaseStatus.SUCCESS.toString()
-
+  @Override protected void ensureLockTtlGreaterThan(String lockName, int ttlSeconds) {
     redisClientDelegate.withCommandsClient({ JedisCommands c ->
       // ensure that the lock exists and has been tt'l corresponding to the `successInterval`
-      assert c.ttl("{korkLock:veryimportantlock}") > 25
+      assert c.ttl("{korkLock:${lockName.toLowerCase()}}") > ttlSeconds
     } as Consumer<JedisCommands>)
+  }
 
-    when:
-    response = redisLockManager.tryReleaseLock(lock, false)
-
-    then:
-    response == LockReleaseStatus.SUCCESS.toString()
-
+  @Override protected void ensureLockTtlLessThanOrEqualTo(String lockName, int ttlSeconds) {
     redisClientDelegate.withCommandsClient({ JedisCommands c ->
       // ensure that the lock exists and has been tt'l corresponding to the `failureInterval`
-      assert c.ttl("{korkLock:veryimportantlock}") <= 25
+      assert c.ttl("{korkLock:${lockName.toLowerCase()}}") <= ttlSeconds
     } as Consumer<JedisCommands>)
   }
 }
