+++ settings.gradle
@@ -20,6 +20,7 @@ include(
   "kork-aws",
   "kork-bom",
   "kork-core",
+  "kork-core-tck",
   "kork-dynomite",
   "kork-exceptions",
   "kork-expressions",
